@@ -1,11 +1,9 @@
 using System;
 using System.Collections.Generic;
 using System.Data;
-using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Text.RegularExpressions;
-using System.Xml.Linq;
 using APSIM.Shared.Documentation;
 using DocumentationGraphPage = APSIM.Shared.Documentation.GraphPage;
 
@@ -118,6 +116,11 @@ public static class DocumentationUtilities
     public static List<ITag> AddHeader(string title, List<ITag> tags) {
 
         List<ITag> newTags = new List<ITag>();
+        if (tags.Count == 0)
+        {
+            return tags;
+        }
+        
         ITag first = tags.First();
         
         if (tags.First().GetType() == typeof(Paragraph))
@@ -131,132 +134,5 @@ public static class DocumentationUtilities
         {
             return tags;
         }
-    }
-
-    /// <summary>
-    /// Get the summary of a member (field, property)
-    /// </summary>
-    /// <param name="member">The member to get the summary for.</param>
-    public static string GetSummary(MemberInfo member)
-    {
-        var fullName = member.ReflectedType + "." + member.Name;
-        if (member is PropertyInfo)
-            return GetSummary(fullName, 'P');
-        else if (member is FieldInfo)
-            return GetSummary(fullName, 'F');
-        else if (member is EventInfo)
-            return GetSummary(fullName, 'E');
-        else if (member is MethodInfo method)
-        {
-            string args = string.Join(",", method.GetParameters().Select(p => p.ParameterType.FullName));
-            args = args.Replace("+", ".");
-            return GetSummary($"{fullName}({args})", 'M');
-        }
-        else
-            throw new ArgumentException($"Unknown argument type {member.GetType().Name}");
-    }
-
-    /// <summary>
-    /// Get the summary of a type removing CRLF.
-    /// </summary>
-    /// <param name="t">The type to get the summary for.</param>
-    public static string GetSummary(Type t)
-    {
-        return GetSummary(t.FullName, 'T');
-    }
-
-    /// <summary>
-    /// Get the summary of a type without removing CRLF.
-    /// </summary>
-    /// <param name="t">The type to get the summary for.</param>
-    public static string GetSummaryRaw(Type t)
-    {
-        return GetSummaryRaw(t.FullName, 'T');
-    }
-
-    /// <summary>
-    /// Get the remarks tag of a type (if it exists).
-    /// </summary>
-    /// <param name="t">The type.</param>
-    public static string GetRemarks(Type t)
-    {
-        return GetRemarks(t.FullName, 'T');
-    }
-
-    /// <summary>
-    /// Get the remarks of a member (field, property) if it exists.
-    /// </summary>
-    /// <param name="member">The member.</param>
-    public static string GetRemarks(MemberInfo member)
-    {
-        var fullName = member.ReflectedType + "." + member.Name;
-        if (member is PropertyInfo)
-            return GetRemarks(fullName, 'P');
-        else if (member is FieldInfo)
-            return GetRemarks(fullName, 'F');
-        else if (member is EventInfo)
-            return GetRemarks(fullName, 'E');
-        else if (member is MethodInfo method)
-        {
-            string args = string.Join(",", method.GetParameters().Select(p => p.ParameterType.FullName));
-            args = args.Replace("+", ".");
-            return GetRemarks($"{fullName}({args})", 'M');
-        }
-        else
-            throw new ArgumentException($"Unknown argument type {member.GetType().Name}");
-    }
-
-    /// <summary>
-    /// Get the summary of a member (class, field, property)
-    /// </summary>
-    /// <param name="path">The path to the member.</param>
-    /// <param name="typeLetter">Type type letter: 'T' for type, 'F' for field, 'P' for property.</param>
-    private static string GetSummary(string path, char typeLetter)
-    {
-        var rawSummary = GetSummaryRaw(path, typeLetter);
-        if (rawSummary != null)
-        {
-            // Need to fix multiline comments - remove newlines and consecutive spaces.
-            return Regex.Replace(rawSummary, @"\n[ \t]+", "\n");
-        }
-        return null;
-    }
-
-    /// <summary>
-    /// Get the summary of a member (class, field, property)
-    /// </summary>
-    /// <param name="path">The path to the member.</param>
-    /// <param name="typeLetter">Type type letter: 'T' for type, 'F' for field, 'P' for property.</param>
-    private static string GetSummaryRaw(string path, char typeLetter)
-    {
-        if (string.IsNullOrEmpty(path))
-            return path;
-
-        path = path.Replace("+", ".");
-
-        //if (summaries.TryGetValue($"{typeLetter}:{path}", out var summary))
-        //    return summary;
-        return null;
-    }
-
-    /// <summary>
-    /// Get the remarks of a member (class, field, property).
-    /// </summary>
-    /// <param name="path">The path to the member.</param>
-    /// <param name="typeLetter">Type letter: 'T' for type, 'F' for field, 'P' for property.</param>
-    /// <returns></returns>
-    private static string GetRemarks(string path, char typeLetter)
-    {
-        if (string.IsNullOrEmpty(path))
-            return path;
-
-        path = path.Replace("+", ".");
-
-        //if (remarks.TryGetValue($"{typeLetter}:{path}", out var remark))
-        //{
-            // Need to fix multiline remarks - trim newlines and consecutive spaces.
-        //    return Regex.Replace(remark, @"\n\s+", "\n");
-        //}
-        return null;
     }
 }
