@@ -10,16 +10,12 @@ using Models.PMF.Interfaces;
 using Models.ForageDigestibility;
 using Newtonsoft.Json;
 using APSIM.Shared.Utilities;
-using Models.Climate;
-using Models.Soils.NutrientPatching;
-using Models.Core.ApsimFile;
-using MathNet.Numerics;
 
 namespace Models.AgPasture
 {
 
     /// <summary>
-    /// A model for cutting pasture / plants and calculating and returning excreta to the
+    ///   A model for cutting pasture / plants and calculating and returning excreta to the
     /// soil based on the biomass cut. If this model is put at the top level of the simulation
     /// all child zones (paddocks) are treated uniformly (e.g. for urine patch modelling)
     /// </summary>
@@ -27,7 +23,6 @@ namespace Models.AgPasture
     [ViewName("UserInterface.Views.PropertyView")]
     [PresenterName("UserInterface.Presenters.PropertyPresenter")]
     [ValidParent(ParentType = typeof(Zone))]
-    [ValidParent(ParentType = typeof(Simulation))]
     public class SimpleGrazing : Model
     {
         [Link] IClock clock = null;
@@ -39,15 +34,9 @@ namespace Models.AgPasture
         private IBooleanFunction expressionFunction;
         private int simpleGrazingFrequency;
         private List<ZoneWithForage> zones;
-        private UrineDungPatches urineDungPatches;
-        private readonly UrineReturnTypes urineReturnType = UrineReturnTypes.FromHarvest;
-        private double[] speciesCutProportions { get; set; }
-
 
         /// <summary>Average potential ME concentration in herbage material (MJ/kg)</summary>
-        private const double potentialMEOfHerbage = 16.0;
-
-        private const double maxEffectiveNConcentration = 3;
+        private const double PotentialMEOfHerbage = 16.0;
 
         /// <summary>Grazing rotation type enum for drop down.</summary>
         public enum GrazingRotationTypeEnum
@@ -69,37 +58,26 @@ namespace Models.AgPasture
         public class UrineDungReturnType : EventArgs
         {
             /// <summary>Grazed dry matter (kg/ha)</summary>
-            public double GrazedDM { get; set; }
+            public double GrazedDM { get; set;  }
 
             /// <summary>N in grazed dry matter (kg/ha).</summary>
-            public double GrazedN { get; set; }
+            public double GrazedN { get; set;  }
 
             /// <summary>Metabolisable energy in grazed dry matter.</summary>
             public double GrazedME { get; set; }
         }
 
-        /// <summary>Urine return patterns.</summary>
-        public enum UrineReturnPatterns
-        {
-            /// <summary>Rotating in order</summary>
-            RotatingInOrder,
-            /// <summary>Not enabled Random</summary>
-            Random,
-            /// <summary>Not enabled Pseudo-random</summary>
-            PseudoRandom
-        }
-
-        /// <summary>Urine return types.</summary>
-        public enum UrineReturnTypes
-        {
-            /// <summary>FromHarvest</summary>
-            FromHarvest,
-            /// <summary>SetMonthly</summary>
-            SetMonthly
-        }
-
         /// <summary>Invoked when a grazing occurs.</summary>
         public event EventHandler Grazed;
+
+        /// <summary>Invoked when urine and dung is to be returned to soil.</summary>
+        /// <remarks>
+        /// This event provides a mechanism for another model to perform a
+        /// urine and dung return to the soil. If no other model subscribes to this
+        /// event then SimpleGrazing will do the return. This mechanism
+        /// allows a urine patch model to work.
+        /// </remarks>
+        public event EventHandler<UrineDungReturnType> DoUrineDungReturn;
 
         ////////////// GUI parameters shown to user //////////////
 
@@ -130,13 +108,13 @@ namespace Models.AgPasture
 
         /// <summary></summary>
         [Separator("Settings for the 'Target Mass' - all values by month from January")]
-        [Description("Target mass of pasture to trigger grazing event (single value or 12 monthly values) (kgDM/ha)")]
+        [Description("Target mass of pasture to trigger grazing event, monthly values (kgDM/ha)")]
         [Units("kgDM/ha")]
         [Display(VisibleCallback = "IsTargetMassTurnedOn")]
         public double[] PreGrazeDMArray { get; set; }
 
         /// <summary></summary>
-        [Description("Residual mass of pasture post grazing (single value or 12 monthly values) (kgDM/ha)")]
+        [Description("Residual mass of pasture post grazing, monthly values (kgDM/ha)")]
         [Units("kgDM/ha")]
         [Display(VisibleCallback = "IsTargetMassTurnedOn")]
         public double[] PostGrazeDMArray { get; set; }
@@ -178,20 +156,13 @@ namespace Models.AgPasture
         public string NoGrazingEndString { get; set; }
 
         /// <summary></summary>
-       // [Separator("Urine and Dung.")]
+        [Separator("Urine and Dung.")]
 
-        // [Description("Use patching to return excreta to the soil?")]
-        // public bool UsePatching { get; set; }
-        [JsonIgnore] public bool UsePatching = false;
-
-        /// <summary>Fraction of defoliated Biomass going to soil</summary>
         [Description("Fraction of defoliated Biomass going to soil. Remainder is exported as animal product or to lanes/camps (0-1).")]
-        [Display(VisibleCallback = "DontUsePatching")]
         public double[] FractionDefoliatedBiomassToSoil { get; set; } = new double[] { 1 };
 
         /// <summary></summary>
         [Description("Fraction of defoliated N going to soil. Remainder is exported as animal product or to lanes/camps (0-1).")]
-        [Display(VisibleCallback = "DontUsePatching")]
         public double[] FractionDefoliatedNToSoil { get; set; }
 
         /// <summary></summary>
@@ -207,7 +178,6 @@ namespace Models.AgPasture
         /// <summary></summary>
         [Description("Depth that urine is added (mm)")]
         [Units("mm")]
-        [Display(VisibleCallback = "DontUsePatching")]
         public double DepthUrineIsAdded { get; set; }
 
         /// <summary></summary>
@@ -219,35 +189,6 @@ namespace Models.AgPasture
         [Description("Send some fraction of the  calculated urine off-paddock - usually this should be zero (0-1)")]
         [Units("0-1")]
         public double SendUrineElsewhere { get; set; }
-
-        // Patching variables.
-
-        /// <summary>Create pseudo patches?</summary>
-        [Description("Should this simulation create pseudo patches? If not then explict zones (slow!) will be created")]
-        [Display(VisibleCallback = "UsePatching")]
-        public bool PseudoPatches { get; set; }
-
-        /// <summary>Number of patches or zones to create.</summary>
-        [Description("How many patches or zones should be created?")]
-        [Display(VisibleCallback = "UsePatching")]
-        public int ZoneCount { get; set; } = 20;
-
-        /// <summary>Urine return pattern.</summary>
-        [Description("Pattern (spatial) of nutrient return")]
-        [Display(VisibleCallback = "UsePatching")]
-        public UrineReturnPatterns UrineReturnPattern { get; set; }
-
-        /// <summary>Seed to use for pseudo random number generator.</summary>
-        [Description("Seed to use for pseudo random number generator")]
-        [Display(VisibleCallback = "IsPseudoRandom")]
-        public int PseudoRandomSeed { get; set; }
-
-        /// <summary>Depth of urine penetration (mm)</summary>
-        [Description("Depth of urine penetration (mm)")]
-        [Display(VisibleCallback = "UsePatching")]
-        public double UrineDepthPenetration { get; set; }
-
-        // End patching variables.
 
         /// <summary></summary>
         [Separator("Plant population modifier")]
@@ -269,31 +210,66 @@ namespace Models.AgPasture
         [Display(VisibleCallback = "IsTramplingTurnedOn")]
         public double PastureConsumedAtMaximumRateOfLitterRemoval { get; set; } = 1200;
 
+        /// <summary></summary>
+        [Separator("Grazing species weighting")]
+        [Description("Optional relative weighting for grazing of forages. Must sum to the number of forages (inc. SurfaceOrganicMatter).")]
+        public double[] SpeciesCutProportions { get; set; }
+
         ////////////// Callbacks to enable/disable GUI parameters //////////////
 
         /// <summary></summary>
-        public bool IsSimpleGrazingTurnedOn => GrazingRotationType == GrazingRotationTypeEnum.SimpleRotation;
+        public bool IsSimpleGrazingTurnedOn
+        {
+            get
+            {
+                return GrazingRotationType == GrazingRotationTypeEnum.SimpleRotation;
+            }
+        }
 
         /// <summary></summary>
-        public bool IsTargetMassTurnedOn => GrazingRotationType == GrazingRotationTypeEnum.TargetMass;
+        public bool IsTargetMassTurnedOn
+        {
+            get
+            {
+                return GrazingRotationType == GrazingRotationTypeEnum.TargetMass;
+            }
+        }
 
         /// <summary></summary>
-        public bool IsNotTimingControlledElsewhere => GrazingRotationType != GrazingRotationTypeEnum.TimingControlledElsewhere;
+        public bool IsNotTimingControlledElsewhere
+        {
+            get
+            {
+                return GrazingRotationType != GrazingRotationTypeEnum.TimingControlledElsewhere;
+            }
+        }
 
         /// <summary></summary>
-        public bool IsFlexibleGrazingTurnedOn => GrazingRotationType == GrazingRotationTypeEnum.Flexible;
+        public bool IsFlexibleGrazingTurnedOn
+        {
+            get
+            {
+                return GrazingRotationType == GrazingRotationTypeEnum.Flexible;
+            }
+        }
 
-        /// <summary>Is CN ratio dung enable?</summary>
-        public bool IsCNRatioDungEnabled => DontUsePatching && FractionExcretedNToDung == null;
+        /// <summary></summary>
+        public bool IsCNRatioDungEnabled
+        {
+            get
+            {
+                return FractionExcretedNToDung == null;
+            }
+        }
 
-        /// <summary>Is pseudo random return pattern selected?</summary>
-        public bool IsPseudoRandom => UrineReturnPattern == UrineReturnPatterns.PseudoRandom;
-
-        /// <summary>Is fraction ExcretedN to dung enabled?</summary>
-        public bool IsFractionExcretedNToDungEnabled => DontUsePatching && (double.IsNaN(CNRatioDung) || CNRatioDung == 0);
-
-        /// <summary>Return true if don't use patching for excreta return.</summary>
-        public bool DontUsePatching => !UsePatching;
+        /// <summary></summary>
+        public bool IsFractionExcretedNToDungEnabled
+        {
+            get
+            {
+                return double.IsNaN(CNRatioDung) || CNRatioDung == 0;
+            }
+        }
 
         /// <summary>
         /// Is maximum rotation length input array enabled in the GUI?
@@ -336,7 +312,7 @@ namespace Models.AgPasture
         /// <summary>N in urine returned to the paddock.</summary>
         [JsonIgnore]
         [Units("kgN/ha")]
-        public double AmountUrineNReturned => urineDungPatches == null ? zones.Sum(z => z.AmountUrineNReturned) : urineDungPatches.AmountUrineNReturned;
+        public double AmountUrineNReturned => zones.Sum(z => z.AmountUrineNReturned);
 
         /// <summary>C in dung returned to the paddock.</summary>
         [JsonIgnore]
@@ -346,7 +322,7 @@ namespace Models.AgPasture
         /// <summary>N in dung returned to the paddock.</summary>
         [JsonIgnore]
         [Units("kgN/ha")]
-        public double AmountDungNReturned => urineDungPatches == null ? zones.Sum(z => z.AmountDungNReturned) : urineDungPatches.AmountDungNReturned;
+        public double AmountDungNReturned => zones.Sum(z => z.AmountDungNReturned);
 
         /// <summary>Mass of herbage just before grazing.</summary>
         [JsonIgnore]
@@ -371,37 +347,10 @@ namespace Models.AgPasture
         /// <summary>Did grazing happen today?</summary>
         [JsonIgnore]
         [Units("0-1")]
-        public bool GrazedToday { get; private set; }
+        public bool GrazedToday{ get; private set; }
 
-        // Patching specific outputs.
-
-        /// <summary>Zone or patch that urine will be applied to</summary>
-        public int ZoneNumForUrine => urineDungPatches == null ? 0 : urineDungPatches.ZoneNumForUrine;
-
-        /// <summary>Number of zones for applying urine</summary>
-        public int NumZonesForUrine => urineDungPatches == null ? 0 : urineDungPatches.NumZonesForUrine;
-
-        /// <summary>Divisor for reporting</summary>
-        public double DivisorForReporting => urineDungPatches == null ? 0 : urineDungPatches.DivisorForReporting;
-
-        /// <summary>Amount of dung carbon returned to soil for whole paddock (kg/ha)</summary>
-        public double AmountDungCReturned => urineDungPatches == null ? 0 : urineDungPatches.AmountDungCReturned;
 
         ////////////// Methods //////////////
-
-        /// <summary>
-        /// Invoked by the infrastructure before the simulation gets created in memory.
-        /// Use this to create patches.
-        /// </summary>
-        public override void OnPreLink()
-        {
-            if (UsePatching)
-            {
-                urineDungPatches = new UrineDungPatches(this, PseudoPatches, ZoneCount, urineReturnType,
-                                                        UrineReturnPattern, PseudoRandomSeed, UrineDepthPenetration, maxEffectiveNConcentration);
-                urineDungPatches.OnPreLink();
-            }
-        }
 
         /// <summary>This method is invoked at the beginning of the simulation.</summary>
         [EventSubscribe("Commencing")]
@@ -421,12 +370,13 @@ namespace Models.AgPasture
                                                                 (z, f) => new ZoneWithForage(z, f.ToList(), areaOfAllZones, summary))
                                                        .ToList();
 
+
             if (GrazingRotationType == GrazingRotationTypeEnum.TargetMass)
             {
-                if (PreGrazeDMArray == null || (PreGrazeDMArray.Length != 1 && PreGrazeDMArray.Length != 12))
-                    throw new Exception("There must be a single value or monthly values specified for 'target mass of pasture to trigger grazing'");
-                if (PostGrazeDMArray == null || (PostGrazeDMArray.Length != 1 && PostGrazeDMArray.Length != 12))
-                    throw new Exception("There must be a single value or monthly values specified for 'residual mass of pasture post grazing'");
+                if (PreGrazeDMArray == null || PreGrazeDMArray.Length != 12)
+                    throw new Exception("There must be 12 values input for the pre-grazing DM");
+                if (PostGrazeDMArray == null || PostGrazeDMArray.Length != 12)
+                    throw new Exception("There must be 12 values input for the post-grazing DM");
             }
             else if (GrazingRotationType == GrazingRotationTypeEnum.Flexible)
             {
@@ -447,8 +397,11 @@ namespace Models.AgPasture
                 numForages = zones.First().NumForages;
             else
                 numForages = zones.Where(z => z.Zone == this.Parent).First().NumForages;
+            if (SpeciesCutProportions == null)
+                SpeciesCutProportions = MathUtilities.CreateArrayOfValues(1.0, numForages);
 
-            speciesCutProportions = MathUtilities.CreateArrayOfValues(1.0, numForages);
+            if (SpeciesCutProportions.Sum() != numForages)
+                throw new Exception("The species cut weightings must add up to the number of species.");
 
             if (SimpleGrazingFrequencyString != null && SimpleGrazingFrequencyString.Equals("end of month", StringComparison.InvariantCultureIgnoreCase))
                 simpleGrazingFrequency = 0;
@@ -456,7 +409,7 @@ namespace Models.AgPasture
                 simpleGrazingFrequency = Convert.ToInt32(SimpleGrazingFrequencyString);
 
             if (FractionDefoliatedNToSoil == null || FractionDefoliatedNToSoil.Length == 0)
-                FractionDefoliatedNToSoil = new double[] { 1 };
+                FractionDefoliatedNToSoil = new double[] { 1};
 
             if (FractionDefoliatedBiomassToSoil == null || FractionDefoliatedBiomassToSoil.Length == 0)
                 FractionDefoliatedBiomassToSoil = new double[] { 0 };
@@ -468,8 +421,6 @@ namespace Models.AgPasture
                       GrazingRotationType == GrazingRotationTypeEnum.Flexible) &&
                       MinimumRotationLengthArray != null)
                 DaysSinceGraze = Convert.ToInt32(MinimumRotationLengthArray[clock.Today.Month - 1]);
-
-            urineDungPatches?.OnStartOfSimulation();
         }
 
         /// <summary>This method is invoked at the beginning of each day to perform management actions.</summary>
@@ -518,7 +469,7 @@ namespace Models.AgPasture
             DaysSinceGraze = 0;
 
             foreach (var zone in zones)
-                zone.RemoveDMFromPlants(residual, speciesCutProportions);
+                zone.RemoveDMFromPlants(residual, SpeciesCutProportions);
 
             DoUrineDungTrampling();
 
@@ -543,20 +494,30 @@ namespace Models.AgPasture
         /// <summary>Add urine to the soil.</summary>
         private void DoUrineDungTrampling()
         {
-            if (UsePatching)
+            if (DoUrineDungReturn == null)
             {
                 if (Parent is Zone)
-                    throw new Exception("To use patches for urine/dung return, SimpleGrazing needs to be at the top level of a simulation.");
-                urineDungPatches.DoUrineDungReturn(zones.First().GrazedN);  // Assumes all zones are harvested the same.
+                {
+                    foreach (var zone in zones)
+                        zone.DoUrineDungTrampling(clock.Today.Month, FractionDefoliatedBiomassToSoil,
+                                                  FractionDefoliatedNToSoil, FractionExcretedNToDung,
+                                                  CNRatioDung, DepthUrineIsAdded, TramplingOn,
+                                                  PastureConsumedAtMaximumRateOfLitterRemoval, MaximumPropLitterMovedToSoil,
+                                                  SendDungElsewhere, SendUrineElsewhere);
+                }
+                else
+                    throw new Exception("Currently, when SimpleGrazing is at the top level of a simulation it must have a SimpleCow sibling present.");
             }
             else
             {
-                foreach (var zone in zones)
-                    zone.DoUrineDungTrampling(clock.Today.Month, FractionDefoliatedBiomassToSoil,
-                                                FractionDefoliatedNToSoil, FractionExcretedNToDung,
-                                                CNRatioDung, DepthUrineIsAdded, TramplingOn,
-                                                PastureConsumedAtMaximumRateOfLitterRemoval, MaximumPropLitterMovedToSoil,
-                                                SendDungElsewhere, SendUrineElsewhere);
+                // Another model (e.g. urine patch) will do the urine return.
+                DoUrineDungReturn.Invoke(this,
+                    new UrineDungReturnType()
+                    {
+                        GrazedDM = GrazedDM,
+                        GrazedN = GrazedN,
+                        GrazedME = GrazedME
+                    });
             }
         }
 
@@ -584,7 +545,7 @@ namespace Models.AgPasture
         /// <returns>True if can graze.</returns>
         private bool TargetMass()
         {
-            residualBiomass = GetValueFromMonthlyArray(clock.Today.Month - 1, PostGrazeDMArray);
+            residualBiomass = PostGrazeDMArray[clock.Today.Month - 1];
 
             // Don't graze if days since last grazing is < minimum
             if (MinimumRotationLengthArray != null && DaysSinceGraze < MinimumRotationLengthArray[clock.Today.Month - 1])
@@ -595,19 +556,8 @@ namespace Models.AgPasture
                 return true;
 
             // Do graze if expression is true
-            return PreGrazeDM > GetValueFromMonthlyArray(clock.Today.Month - 1, PreGrazeDMArray);
-        }
-
-        /// <summary>
-        /// Helper function to return a monthly value from an array that may have
-        /// number in it or 12 numbers.
-        /// </summary>
-        /// <param name="monthIndex">The index</param>
-        /// <param name="array">The array</param>
-        /// <returns></returns>
-        private double GetValueFromMonthlyArray(int monthIndex, double[] array)
-        {
-            return array.Length == 1 ? array[0] : array[monthIndex];
+            // this was return PreGrazeHarvestableDM > PreGrazeDMArray[clock.Today.Month - 1]; but this assessment shoudl be against total DM
+            return PreGrazeDM > PreGrazeDMArray[clock.Today.Month - 1];
         }
 
         /// <summary>Calculate whether a target mass and length rotation can graze today.</summary>
@@ -626,8 +576,9 @@ namespace Models.AgPasture
 
             // Do graze if expression is true
             else
-                return expressionFunction.Value();
+                return expressionFunction.Value();  
         }
+
 
         private class ZoneWithForage
         {
@@ -748,7 +699,7 @@ namespace Models.AgPasture
                     }
 
                     // If a fraction consumable was specified in the forages component by the user then the above calculated
-                    // removeAmount might be > consumable amount. Constrain the removeAmount to the consumable
+                    // removeAmount might be > consumable amount. Constrain the removeAmount to the consumable 
                     // amount so that we don't get an exception thrown in ModelWithDigestibleBiomass.RemoveBiomass method
                     removeAmount = Math.Min(removeAmount, totalHarvestableWt);
 
@@ -761,7 +712,7 @@ namespace Models.AgPasture
                         {
                             var grazed = forages[i].RemoveBiomass(amountToRemove: amountToRemove);
                             double grazedDigestibility = grazed.Digestibility;
-                            var grazedMetabolisableEnergy = potentialMEOfHerbage * grazedDigestibility;
+                            var grazedMetabolisableEnergy = PotentialMEOfHerbage * grazedDigestibility;
 
                             grazedDM += grazed.Wt;  // kg/ha
                             grazedN += grazed.N;    // kg/ha
@@ -798,7 +749,7 @@ namespace Models.AgPasture
                                     double sendDungElsewhere,
                                     double sendUrineElsewhere)
             {
-                var urineDung = UrineDungReturn.CalculateUrineDungReturn(grazedForages,
+                var urineDung = UrineDungReturn.CalculateUrineDungReturn(grazedForages,  
                                                                          GetValueFromMonthArray(fractionDefoliatedBiomassToSoil, month),
                                                                          GetValueFromMonthArray(fractionDefoliatedNToSoil, month),
                                                                          GetValueFromMonthArray(fractionExcretedNToDung, month),
@@ -835,360 +786,6 @@ namespace Models.AgPasture
                     return arr[0];
                 else
                     return arr[month - 1];
-            }
-        }
-
-        // **************************************************************************************
-        // Patching code
-        // **************************************************************************************
-
-        /// <summary>
-        /// Encapsulates urine patch functionality.
-        /// </summary>
-        public class UrineDungPatches
-        {
-            private readonly SimpleGrazing simpleGrazing;
-            private readonly bool pseudoPatches;
-            private double[] monthlyUrineNAmt;                 // breaks the N balance but useful for testing
-            private double[] urineDepthPenetrationArray;
-            private Random pseudoRandom;
-            private int pseudoRandomSeed;
-            private readonly ISummary summary;
-            private readonly Clock clock;
-            private readonly Physical physical;
-
-            // User properties.
-
-            /// <summary>Number of patches or zones to create.</summary>
-            private readonly int zoneCount;
-
-            /// <summary>Urine return type</summary>
-            private readonly UrineReturnTypes urineReturnType;
-
-            /// <summary>Urine return pattern.</summary>
-            private readonly UrineReturnPatterns urineReturnPattern;
-
-            /// <summary>Depth of urine penetration (mm)</summary>
-            private readonly double urineDepthPenetration;
-
-            /// <summary>Maximum effective NO3-N or NH4-N concentration</summary>
-            private readonly double maxEffectiveNConcentration;
-
-            // Outputs.
-
-            /// <summary>Zone or patch that urine will be applied to</summary>
-            public int ZoneNumForUrine { get; private set; }
-
-            /// <summary>Number of zones for applying urine</summary>
-            public int NumZonesForUrine { get; private set; }
-
-            /// <summary>Divisor for reporting</summary>
-            public double DivisorForReporting { get; private set; }
-
-            /// <summary>Amount of urine returned to soil for whole paddock (kg/ha)</summary>
-            public double AmountUrineNReturned { get; private set; }
-
-            /// <summary>Amount of dung nitrogen returned to soil for whole paddock (kg/ha)</summary>
-            public double AmountDungNReturned { get; private set; }
-
-            /// <summary>Amount of dung carbon returned to soil for whole paddock (kg/ha)</summary>
-            public double AmountDungCReturned { get; private set; }
-
-            /// <summary>
-            /// Constructor
-            /// </summary>
-            /// <param name="simpleGrazing">Parent SimpleGrazing model</param>
-            /// <param name="pseudoPatches">Use pseudo patches?</param>
-            /// <param name="zoneCount"></param>
-            /// <param name="urineReturnType"></param>
-            /// <param name="urineReturnPattern"></param>
-            /// <param name="pseudoRandomSeed"></param>
-            /// <param name="urineDepthPenetration"></param>
-            /// <param name="maxEffectiveNConcentration"></param>
-            public UrineDungPatches(SimpleGrazing simpleGrazing, bool pseudoPatches,
-                                    int zoneCount,
-                                    UrineReturnTypes urineReturnType,
-                                    UrineReturnPatterns urineReturnPattern,
-                                    int pseudoRandomSeed,
-                                    double urineDepthPenetration,
-                                    double maxEffectiveNConcentration)
-            {
-                this.simpleGrazing = simpleGrazing;
-                this.pseudoPatches = pseudoPatches;
-                this.zoneCount = zoneCount;
-                this.urineReturnType = urineReturnType;
-                this.urineReturnPattern = urineReturnPattern;
-                this.pseudoRandomSeed = pseudoRandomSeed;
-                this.urineDepthPenetration = urineDepthPenetration;
-                this.maxEffectiveNConcentration = maxEffectiveNConcentration;
-                summary = simpleGrazing.FindInScope<ISummary>();
-                clock = simpleGrazing.FindInScope<Clock>();
-                physical = simpleGrazing.FindInScope<Physical>();
-            }
-
-            /// <summary>
-            /// Invoked by the infrastructure before the simulation gets created in memory.
-            /// Use this to create patches.
-            /// </summary>
-            public void OnPreLink()
-            {
-                var simulation = simpleGrazing.FindAncestor<Simulation>() as Simulation;
-                var zone = simulation.FindChild<Zone>();
-
-                if (zoneCount == 0)
-                    throw new Exception("Number of patches/zones in urine patches is zero.");
-
-                if (pseudoPatches)
-                {
-                    zone.Area = 1.0;
-
-                    var patchManager = simpleGrazing.FindInScope<NutrientPatchManager>();
-                    if (patchManager == null)
-                        throw new Exception("Cannot find NutrientPatchManager");
-                    var soilPhysical = simpleGrazing.FindInScope<Physical>();
-                    if (patchManager == null)
-                        throw new Exception("Cannot find Physical");
-
-                    double[] ArrayForMaxEffConc = new double[soilPhysical.Thickness.Length];
-                    for (int i = 0; i <= (soilPhysical.Thickness.Length - 1); i++)
-                        ArrayForMaxEffConc[i] = maxEffectiveNConcentration;
-
-                    patchManager.MaximumNO3AvailableToPlants = ArrayForMaxEffConc;
-                    patchManager.MaximumNH4AvailableToPlants = ArrayForMaxEffConc;
-
-                    patchManager.NPartitionApproach = PartitionApproachEnum.BasedOnConcentrationAndDelta;
-                    patchManager.AutoAmalgamationApproach = AutoAmalgamationApproachEnum.None;
-                    patchManager.basePatchApproach = BaseApproachEnum.IDBased;
-                    patchManager.AllowPatchAmalgamationByAge = false;
-                    patchManager.PatchAgeForForcedMerge = 1000000.0;  // ie don't merge
-
-                    int[] PatchToAddTo = new int[1];  //need an array variable for this
-                    string[] PatchNmToAddTo = new string[1];
-                    int nPatchesAdded = 0;
-                    double NewArea = 1.0 / zoneCount;
-
-                    while (nPatchesAdded < zoneCount - 1)
-                    {
-                        AddSoilCNPatchType NewPatch = new AddSoilCNPatchType();
-                        NewPatch.DepositionType = DepositionTypeEnum.ToNewPatch;
-                        NewPatch.AreaFraction = NewArea;
-                        PatchToAddTo[0] = 0;
-                        PatchNmToAddTo[0] = "0";
-                        NewPatch.AffectedPatches_id = PatchToAddTo;
-                        NewPatch.AffectedPatches_nm = PatchNmToAddTo;
-                        NewPatch.SuppressMessages = false;
-                        patchManager.Add(NewPatch);
-                        nPatchesAdded += 1;
-                    }
-
-                }
-                else //(!PseudoPatches)  // so now this is zones - possibly multiple zones
-                {
-                    zone.Area = 1.0 / zoneCount;  // and then this will apply to all the new zones
-                    for (int i = 0; i < zoneCount - 1; i++)
-                    {
-                        var newZone = Apsim.Clone(zone);
-                        Structure.Add(newZone, simulation);
-                    }
-                }
-            }
-
-            /// <summary>Invoked at start of simulation.</summary>
-            public void OnStartOfSimulation()
-            {
-                if (!pseudoPatches)
-                    summary.WriteMessage(simpleGrazing, "Created " + zoneCount + " identical zones, each of area " + (1.0 / zoneCount) + " ha", MessageType.Diagnostic);
-
-                summary.WriteMessage(simpleGrazing, "Initialising the ZoneManager for grazing, urine return and reporting", MessageType.Diagnostic);
-
-                pseudoRandom = new Random(pseudoRandomSeed);  // sets a constant seed value
-
-                if (pseudoPatches)
-                    DivisorForReporting = 1.0;
-                else
-                    DivisorForReporting = zoneCount;
-
-                monthlyUrineNAmt = new double[] { 24, 19, 17, 12, 8, 5, 5, 10, 16, 19, 23, 25 }; //This is to get a pattern of return that varies with month but removes the variation that might be caused by small changes in herbage growth
-                //MonthlyUrineNAmt = new double[] { 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25, 25 }; //This is to get a pattern of return that varies with month but removes the variation that might be caused by small changes in herbage growth
-                //MonthlyUrineNAmt = new double[] { 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0 }; //This is to get a pattern of return that varies with month but removes the variation that might be caused by small changes in herbage growth
-
-                if (pseudoPatches)
-                {
-                    var patchManager = simpleGrazing.FindInScope<NutrientPatchManager>();
-
-                    //var patchManager = FindInScope<NutrientPatchManager>();
-                    summary.WriteMessage(simpleGrazing, patchManager.NumPatches.ToString() + " pseudopatches have been created", MessageType.Diagnostic);
-                }
-                else
-                {
-                    var simulation = simpleGrazing.FindAncestor<Simulation>();
-                    var physical = simpleGrazing.FindInScope<IPhysical>();
-                    double[] arrayForMaxEffConc = Enumerable.Repeat(maxEffectiveNConcentration, physical.Thickness.Length).ToArray();
-                    foreach (Zone zone in simulation.FindAllInScope<Zone>())
-                    {
-                        foreach (var patchManager in zone.FindAllInScope<NutrientPatchManager>())
-                        {
-                            patchManager.MaximumNO3AvailableToPlants = arrayForMaxEffConc;
-                            patchManager.MaximumNH4AvailableToPlants = arrayForMaxEffConc;
-                        }
-                    }
-                }
-
-                summary.WriteMessage(simpleGrazing, "Finished initialising the Manager for grazing, urine return and reporting", MessageType.Diagnostic);
-
-                NumZonesForUrine = 1;  // in the future this might be > 1
-                ZoneNumForUrine = -1;  // this will be incremented to 0 (first zone) below
-
-                UrinePenetration();
-            }
-
-            /// <summary>Invoked at DoManagement.</summary>
-            public void DoUrineDungReturn(double harvestedN)
-            {
-                if (urineReturnType == UrineReturnTypes.FromHarvest)
-                {
-                    AmountUrineNReturned = harvestedN * 0.50;  //
-                    AmountDungNReturned = harvestedN * 0.35;  //
-                    AmountDungCReturned = AmountDungNReturned * 20;
-                }
-                else if (urineReturnType == UrineReturnTypes.SetMonthly)
-                {
-                    AmountUrineNReturned = monthlyUrineNAmt[clock.Today.Month - 1];   //  hardcoded as an input
-                    AmountDungNReturned = AmountUrineNReturned / 0.50 * 0.35;  //
-                    AmountDungCReturned = AmountDungNReturned * 20;
-                }
-                summary.WriteMessage(simpleGrazing, "The amount of urine N to be returned to the whole paddock is " + AmountUrineNReturned, MessageType.Diagnostic);
-
-                DoUrineReturn();
-
-                DoTramplingAndDungReturn();
-
-                summary.WriteMessage(simpleGrazing, "Finished Grazing", MessageType.Diagnostic);
-            }
-
-            /// <summary>Invoked to do trampling and dung return.</summary>
-            private void DoTramplingAndDungReturn()
-            {
-                // Note that dung is assumed to be spread uniformly over the paddock (patches or sones).
-                // There is no need to bring zone area into the calculations here but zone area must be included for variables reported FROM the zone to the upper level
-
-                int i = -1;  // patch or paddock counter
-                foreach (Zone zone in simpleGrazing.FindAllInScope<Zone>())
-                {
-                    i += 1;
-                    SurfaceOrganicMatter surfaceOM = zone.FindInScope<SurfaceOrganicMatter>() as SurfaceOrganicMatter;
-
-                    // do some trampling of litter
-                    // accelerate the movement of surface litter into the soil - do this before the dung is added
-                    double temp = surfaceOM.Wt * 0.1;
-
-                    surfaceOM.Incorporate(fraction: (double)0.1, depth: (double)100.0, doOutput: true);
-
-                    summary.WriteMessage(simpleGrazing, "For patch " + i + " the amount of litter trampled was " + temp + " and the remaining litter is " + (surfaceOM.Wt), MessageType.Diagnostic);
-
-                    // move the dung to litter
-                    AddFaecesType dung = new()
-                    {
-                        OMWeight = AmountDungCReturned / 0.4,  //assume dung C is 40% of OM
-                        OMN = AmountDungNReturned
-                    };
-                    surfaceOM.AddFaeces(dung);
-                    summary.WriteMessage(simpleGrazing, "For patch " + i + " the amount of dung DM added to the litter was " + (AmountDungCReturned / 0.4) + " and the amount of N added in the dung was " + (AmountDungNReturned), MessageType.Diagnostic);
-
-                }
-            }
-
-            /// <summary>Invoked to do urine return</summary>
-            private void DoUrineReturn()
-            {
-                GetZoneForUrineReturn();
-
-                summary.WriteMessage(simpleGrazing, "The Zone for urine return is " + ZoneNumForUrine, MessageType.Diagnostic);
-
-                if (!pseudoPatches)
-                {
-                    Zone zone = simpleGrazing.FindAllInScope<Zone>().ToArray()[ZoneNumForUrine];
-                    Fertiliser thisFert = zone.FindInScope<Fertiliser>() as Fertiliser;
-
-                    thisFert.Apply(amount: AmountUrineNReturned * zoneCount,
-                            type: Fertiliser.Types.UreaN,
-                            depthTop: 0.0,
-                            depthBottom: urineDepthPenetration,
-                            doOutput: true);
-
-                    summary.WriteMessage(simpleGrazing, AmountUrineNReturned + " urine N added to Zone " + ZoneNumForUrine + ", the local load was " + AmountUrineNReturned / zone.Area + " kg N /ha", MessageType.Diagnostic);
-                }
-                else // PseudoPatches
-                {
-                    int[] PatchToAddTo = new int[1];  //because need an array variable for this
-                    string[] PatchNmToAddTo = new string[0];  //need an array variable for this
-                    double[] UreaToAdd = new double[physical.Thickness.Length];
-
-                    for (int ii = 0; ii <= (physical.Thickness.Length - 1); ii++)
-                        UreaToAdd[ii] = urineDepthPenetrationArray[ii] * AmountUrineNReturned * zoneCount;
-
-                    // needed??   UreaReturned += AmountFertNReturned;
-
-                    AddSoilCNPatchType CurrentPatch = new();
-                    CurrentPatch.Sender = "manager";
-                    CurrentPatch.DepositionType = DepositionTypeEnum.ToSpecificPatch;
-                    PatchToAddTo[0] = ZoneNumForUrine;
-                    CurrentPatch.AffectedPatches_id = PatchToAddTo;
-                    CurrentPatch.AffectedPatches_nm = PatchNmToAddTo;
-                    CurrentPatch.Urea = UreaToAdd;
-
-                    var patchManager = simpleGrazing.FindInScope<NutrientPatchManager>();
-
-                    summary.WriteMessage(simpleGrazing, "Patch MinN prior to urine return: " + patchManager.MineralNEachPatch[ZoneNumForUrine], MessageType.Diagnostic);
-                    patchManager.Add(CurrentPatch);
-                    summary.WriteMessage(simpleGrazing, "Patch MinN after urine return: " + patchManager.MineralNEachPatch[ZoneNumForUrine], MessageType.Diagnostic);
-                }
-            }
-
-            /// <summary>Determine and return the zone for urine return.</summary>
-            private void GetZoneForUrineReturn()
-            {
-                if (urineReturnPattern == UrineReturnPatterns.RotatingInOrder)
-                {
-                    ZoneNumForUrine += 1;  //increment the zone number - it was initialised at -1. NOTE, ZoneNumForUrine is used for both zones and patches
-                    if (ZoneNumForUrine >= zoneCount)
-                        ZoneNumForUrine = 0;  // but reset back to the first patch if needed
-                }
-                else if (urineReturnPattern == UrineReturnPatterns.Random)
-                {
-                    Random rnd = new Random();
-                    ZoneNumForUrine = rnd.Next(0, zoneCount); // in C# the maximum value (ZoneCount) will not be selected
-                }
-                else if (urineReturnPattern == UrineReturnPatterns.PseudoRandom)
-                {
-                    ZoneNumForUrine = pseudoRandom.Next(0, zoneCount); // in C# the maximum value (ZoneCount) will not be selected
-                }
-                else
-                    throw new Exception("UrineResturnPattern not recognised");
-
-                summary.WriteMessage(simpleGrazing, "The next zone/patch for urine return is " + ZoneNumForUrine, MessageType.Diagnostic);
-            }
-
-            /// <summary>Calculate the urine penetration array.</summary>
-            private void UrinePenetration()
-            {
-                // note this assumes that all the paddocks are the same
-                double tempDepth = 0.0;
-                urineDepthPenetrationArray = new double[physical.Thickness.Length];
-                for (int i = 0; i <= (physical.Thickness.Length - 1); i++)
-                {
-                    tempDepth += physical.Thickness[i];
-                    if (tempDepth <= urineDepthPenetration)
-                    {
-                        urineDepthPenetrationArray[i] = physical.Thickness[i] / urineDepthPenetration;
-                    }
-                    else
-                    {
-                        urineDepthPenetrationArray[i] = (urineDepthPenetration - (tempDepth - physical.Thickness[i])) / (tempDepth - (tempDepth - physical.Thickness[i])) * physical.Thickness[i] / urineDepthPenetration;
-                        urineDepthPenetrationArray[i] = Math.Max(0.0, Math.Min(1.0, urineDepthPenetrationArray[i]));
-                    }
-                    summary.WriteMessage(simpleGrazing, "The proportion of urine applied to the " + i + "th layer will be " + urineDepthPenetrationArray[i], MessageType.Diagnostic);
-                }
             }
         }
     }
